See extended description
Added glow-in-dark effect to tinker table
Included sprites from SebK and adjusted gui to use them
Changed module architecture (again) to be more scaleable in the future
and more closely resemble existing energynet designs (mostly UE)
A few packet handling tweaks

+++ machinemuse/powersuits/gui/ClickableItem.java
@@ -50,10 +50,9 @@ public class ClickableItem extends Clickable {
 	 */
 	@Override
 	public void draw(RenderEngine engine, MuseGui gui) {
-
 		MuseRenderer.drawItemAt(
-				(int) gui.absX(getPosition().x()) - offsetx,
-				(int) gui.absY(getPosition().y()) - offsety,
+				gui.absX(getPosition().x()) - offsetx,
+				gui.absY(getPosition().y()) - offsety,
 				gui, item);
 	}
 }
